+++ data.json
@@ -51,7 +51,7 @@
       "writer": "healthydog",
       "ingredients": ["연어", "감자 가루", "달걀", "오메가 3 보충제"],
       "description": "고급 간식 연어 스틱 레시피로 건강한 영양소를 공급해보세요.",
-      "main_image_url": "http://localhost:3000/images/recipe02.jpg",
+      "main_image_url": "http://localhost:3000/images/recipe05.jpg",
       "likes": 0
     },
     {
@@ -62,7 +62,7 @@
       "writer": "fitcook",
       "ingredients": ["닭가슴살", "양파", "마늘", "간장"],
       "description": "건강한 닭가슴살 구이 레시피로 단백질 공급!",
-      "main_image_url": "http://localhost:3000/images/recipe03.jpg",
+      "main_image_url": "http://localhost:3000/images/recipe06.jpg",
       "likes": 0
     },
     {
@@ -73,7 +73,7 @@
       "writer": "veganlover",
       "ingredients": ["시금치", "고구마", "찹쌀 가루", "소금"],
       "description": "비건 레시피로 만드는 시금치 고구마 볼!",
-      "main_image_url": "http://localhost:3000/images/recipe04.jpg",
+      "main_image_url": "http://localhost:3000/images/recipe07.jpg",
       "likes": 0
     },
     {
@@ -84,7 +84,7 @@
       "writer": "healthylife",
       "ingredients": ["바나나", "요거트", "우유", "꿀"],
       "description": "간단하게 만드는 바나나 스무디 레시피!",
-      "main_image_url": "http://localhost:3000/images/recipe05.jpg",
+      "main_image_url": "http://localhost:3000/images/recipe08.jpg",
       "likes": 0
     },
     {
@@ -95,7 +95,7 @@
       "writer": "healthydog",
       "ingredients": ["사과", "치즈", "코코넛 가루", "계피"],
       "description": "과일과 치즈가 어우러진 맛있는 간식 사과 치즈 볼!",
-      "main_image_url": "http://localhost:3000/images/recipe06.jpg",
+      "main_image_url": "http://localhost:3000/images/recipe09.jpg",
       "likes": 0
     },
     {
@@ -106,8 +106,52 @@
       "writer": "fitcook",
       "ingredients": ["양파", "밀가루", "우유", "파우더"],
       "description": "간단하게 만드는 양파링 레시피!",
-      "main_image_url": "http://localhost:3000/images/recipe07.jpg",
+      "main_image_url": "http://localhost:3000/images/recipe10.jpg",
       "likes": 0
+    },
+    {
+      "num": 11,
+      "title": "맛있는 닭가슴살 베이크",
+      "level": "중급",
+      "time": "35분",
+      "writer": "foodlover",
+      "ingredients": ["닭가슴살", "감자", "양파", "로즈마리"],
+      "description": "오븐에서 굽는 맛있는 닭가슴살 베이크 레시피입니다.",
+      "main_image_url": "http://localhost:3000/images/recipe11.jpg",
+      "likes": 11
+    },
+    {
+      "num": 12,
+      "title": "간편한 치킨 너겟",
+      "level": "초급",
+      "time": "25분",
+      "writer": "quickcook",
+      "ingredients": ["닭가슴살", "빵가루", "밀가루", "달걀"],
+      "description": "간편하게 만드는 치킨 너겟 레시피입니다.",
+      "main_image_url": "http://localhost:3000/images/recipe12.jpg",
+      "likes": 5
+    },
+    {
+      "num": 13,
+      "title": "간편한 오트밀 머핀",
+      "level": "초급",
+      "time": "30분",
+      "writer": "oatmeallover",
+      "ingredients": ["오트밀", "바나나", "요거트", "계란"],
+      "description": "간편하게 만드는 오트밀 머핀 레시피입니다.",
+      "main_image_url": "http://localhost:3000/images/recipe05.jpg",
+      "likes": 15
+    },
+    {
+      "num": 14,
+      "title": "사과 바나나 요거트 스무디",
+      "level": "초급",
+      "time": "10분",
+      "writer": "healthysmoothie",
+      "ingredients": ["사과", "바나나", "요거트", "우유"],
+      "description": "간단하게 만드는 사과 바나나 요거트 스무디 레시피입니다.",
+      "main_image_url": "http://localhost:3000/images/recipe02.jpg",
+      "likes": 8
     }
   ],
  
@@ -196,6 +240,34 @@
           }
     ]
 
-  }
+  },
+  "partyList": [ 
+    {
+      "title": "뽀삐의 생일파티🐶🎉", 
+      "main_image_url": "http://localhost:3000/images/party_img01.jpg", 
+      "link": "/category/etc"  
+     },
+     {
+      "title": "우리집 멍뭉이의 생일파티💛", 
+      "main_image_url": "http://localhost:3000/images/party_img02.jpg", 
+      "link": "/category/etc"  
+     },
+     {
+      "title": "둥이들 100일파티😻", 
+      "main_image_url": "http://localhost:3000/images/party_img03.jpg", 
+      "link": "/category/etc"  
+     },
+     {
+      "title": "초코 1000일 기념❤️❤️", 
+      "main_image_url": "http://localhost:3000/images/party_img04.jpg", 
+      "link": "/category/etc"  
+     },
+     {
+      "title": "햄냥이의 파티에 초대합니다!🎉🎉", 
+      "main_image_url": "http://localhost:3000/images/party_img05.png", 
+      "link": "/category/etc"  
+     }
+    
+  ]
 
 }
