+++ src/App.js
@@ -3,10 +3,9 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 import MainPage from './pages/MainPage/MainPage.js';
 import Header from './component/Layout/Header/Header.js';
 import Footer from './component/Layout/Footer/Footer.js';
-import RecipeBoard from './pages/RecipeBoard.js';
 import RecipeDetail from './pages/Recipe/RecipeDetail.js'
-import Notice from './pages/Notice/NoticeDetail.js'
-import Recipe from './pages/RecipeBoard.js'
+import Notice from './pages/Notice/NoticeBoard.js'
+import Recipe from './pages/Recipe/RecipeBoard.js'
 import Party from './pages/Party/PartyDetail.js'
 import Login from './pages/Login/Login.js'
 import Signup from './pages/Signup/Signup.js'
