+++ src/component/Profile/Profile.css
@@ -1,8 +1,7 @@
 .profile-container {
-  width: 350px;
-  height: 500px;
+  width: 300px;
+  height: 400px;
   margin-top: 60px;
-  margin-left: 250px;
   display: flex;
   flex-direction: column;
   text-align: center;
@@ -71,7 +70,7 @@
 }
 
 .profile-info hr {
-  width: 300px;
+  width: 230px;
   margin-top: 20px;
 }
