+++ src/pages/MainPage/MainPage.js
@@ -1,11 +1,31 @@
-import React ,{ useState} from 'react'
+import React ,{ useState, useEffect} from 'react'
 import './MainPage.css'
 import Banner from '../../component/Banner/Banner';
 import CardList from '../../component/CardList/CardList';
+import axios from 'axios';
 
 function MainPage() {
 
-  const [cards, setCards] = useState([]);
+  const [recipeCards, setRecipeCards] = useState([]);
+    const [partyCards, setPartyCards] = useState([]);
+
+    useEffect(() => {
+        axios.get('http://localhost:5000/recipes')
+            .then(response => {
+                setRecipeCards(response.data.recipeCards);
+            })
+            .catch(error => {
+                console.error('Error fetching recipe cards:', error);
+            });
+
+        axios.get('http://localhost:5000/partyList')
+            .then(response => {
+                setPartyCards(response.data.partyCards);
+            })
+            .catch(error => {
+                console.error('Error fetching chef cards:', error);
+            });
+    }, []);
 
   return (
     <div className='container'>
@@ -14,12 +34,13 @@ function MainPage() {
         </div>
 
         <div className='board-title'>지금 <span style={{color:"#ff6a10"}} >핫🔥한 </span>레시피</div>
-        <CardList cards={cards.slice(4, 8)} showTitle={false} />
+        <CardList cards={recipeCards} showTitle={false} />
         
         <div className='board-title'>Today <span style={{ color: "#ff6a10" }} >레시피 </span></div>
-        <CardList cards={cards.slice(4, 8)} showTitle={false} />
+          <CardList cards={recipeCards} showTitle={false} />
 
         <div className='board-title'>오늘의 🎉<span style={{ color: "#ff6a10" }} >주인공 </span></div>
+        <CardList cards={partyCards} showTitle={false} />
 
     </div>
   )
