+++ src/pages/Mypage/Mypage.css
@@ -1,4 +1,5 @@
-.mypage-content {
+.myPage-content {
   display: flex;
   justify-content: space-between;
 }
+
