+++ src/pages/Notice/NoticeBoard.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const NoticeBoard = () => {
+    return (
+        <div className='container'>
+            <div className='board'></div>
+            
+        </div>
+    );
+};
+
+export default NoticeBoard;
