+++ src/pages/Signup/Signup.css
@@ -3,7 +3,6 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-  margin : 150px 0 200px;
 }
 
 .signup-container h1 {
@@ -11,6 +10,7 @@
   align-items: center;
   justify-content: flex-start;
   width: 370px;
+  margin-top: 200px;
   margin-bottom: 3px;
   font-size: 24px;
   font-weight: 700;
@@ -38,13 +38,14 @@
 }
 
 .signup-form input {
-  width: 370px;
+  width: 360px;
   height: 40px;
   border: 1px solid #d3d3d3;
   border-radius: 5px;
   margin-top: 3px;
   margin-bottom: 15px;
   font-size: 15px;
+  padding-left: 10px;
 }
 
 .emailCheck-button {
@@ -83,8 +84,8 @@
 }
 
 .signup-pets {
-  display: flex;
-  flex-direction: column;
+  display: grid;
+  grid-template-columns: 100px 100px 100px;
   align-items: center;
   margin-top: 20px;
 }
@@ -103,13 +104,14 @@
 
 .signup-agree {
   display: flex;
+  justify-content: center;
   width: 370px;
-  height: 180px;
+  height: 220px;
   flex-direction: column;
   align-items: center;
   border: 1px solid #ccc;
   border-radius: 5px;
-  margin-top: 20px;
+  margin: 20px 0;
 }
 
 .signup-agree label {
@@ -122,6 +124,7 @@
   margin-right: 8px;
   width: 13px;
   height: 13px;
+  margin-top: 16px;
 }
 
 .signup-button {
